refactor(server)Renamed arena states
